+++ BookGen.Launch/Properties/Resources.Designer.cs
@@ -61,6 +61,27 @@ namespace BookGen.Launch.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to # 2021.09
+        ///* New: XML documentation generator reworked. Now uses XmlDocMarkdown
+        ///* New: Syntax rendering reworked, improved
+        ///* New: Rewrote launcher (again), Has more features
+        ///	* New layout
+        ///	* Options to start VS code, File explorer, Preview and bookgen shell for folder
+        ///	* Filtering for recent folders
+        ///	* Integrated changelog
+        ///	* Will be used as base point for future installers
+        ///* Fix: Improved XHTML compatibility for epub
+        ///* Fix: Fixed a memory leak in the generator
+        ///* Fix: Various code quality fixes
+        ///* [rest of string was truncated]&quot;;.
+        /// </summary>
+        internal static string Changelog {
+            get {
+                return ResourceManager.GetString("Changelog", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Do you want to clear the recent files list?.
         /// </summary>
         internal static string ClearRecentList {
@@ -84,6 +105,15 @@ namespace BookGen.Launch.Properties {
         internal static string FolderselectDescription {
             get {
                 return ResourceManager.GetString("FolderselectDescription", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Install tool to the PATH environment variable?.
+        /// </summary>
+        internal static string InstallToPathVar {
+            get {
+                return ResourceManager.GetString("InstallToPathVar", resourceCulture);
             }
         }